move input upload to own start section
