+++ src/js/components/lists/articles/state/types.js
@@ -1,5 +1,1 @@
 export const RECIEVE_SINGLE_ARTICLE = 'RECIEVE_SINGLE_ARTICLE';
-
-export const POST_ARTICLE_UPLOAD = 'START_ARTICLE_UPLOAD'
-export const SUCCESS_ARTICLE_UPLOAD = 'START_ARTICLE_UPLOAD'
-export const WAIT_ARTICLE_UPLOAD = 'START_ARTICLE_UPLOAD'
