+++ src/js/components/sections/input/state/types.js
@@ -1,0 +1,3 @@
+export const POST_ARTICLE_UPLOAD = 'START_ARTICLE_UPLOAD'
+export const SUCCESS_ARTICLE_UPLOAD = 'START_ARTICLE_UPLOAD'
+export const WAIT_ARTICLE_UPLOAD = 'START_ARTICLE_UPLOAD'
